requirements.txt
flask
flask-socketio
mysql-connector-python
openai
requests
